+++ app/src/Helpers/name_helpers.ts
@@ -1,6 +1,7 @@
 import { memoize } from 'lodash';
 import type { CollectionID } from '@exweiv/weiv-data'
 import { kaptanLogar } from '../Errors/error_manager';
+import { getWeivDataConfigs } from '../Config/weiv_data_config';
 
 export const splitCollectionId = memoize(splitCollectionIdMain);
 function splitCollectionIdMain(collectionId: CollectionID): { dbName: string, collectionName: string } {
@@ -9,9 +10,11 @@ function splitCollectionIdMain(collectionId: CollectionID): { dbName: string, co
     }
 
     const [dbName, collectionName] = collectionId.split('/');
+    const { defaultDatabaseName } = getWeivDataConfigs();
 
     if (!dbName || !collectionName) {
-        return { dbName: "ExWeiv", collectionName: dbName };
+        // When no dbname passed first value is the collection name so default db name is used here.
+        return { dbName: defaultDatabaseName || "ExWeiv", collectionName: dbName };
     }
 
     return { dbName, collectionName };